DommeImage Overhaul Reworked the Function to search a DommeImage, that is tagged with the given Domme Tags.Now it is possible to Exclude Tags from Search. The Tag-Order, case and count doesn't matter.
You want to show a butt without feet, you can enter "Ass, NotFeet".
You want to show a closeup face without boobs: "Face, NotBoobs, Closeup"

If there is no image found for the specified Tags, the Tags will be altered and searched again:

The order of alternation is:
1. Remove: Furniture, SexToy, Tattoo
2. Remove: Closeup, Sideview
3. Change: Naked -> GarmentCovering
4. Change: GarmentCovering -> HalfDressed
5. Change: HalfDressed -> FullDressed
6. Change: HandsCovering -> GarmentCovering
7. Remove: Excluded Tags from the BaseTags
8-12: Same as 1-6 without Excluded tags. If there are no excluded tags this will be skipped.
13. Change: FullDressed -> HalfDressed
14. Change: HalfDressed -> GarmentCovering

Before each step there is a check, if it could alter the result. If it won't the Step is skipped.